Generalize info fetching from fork children
In principle I only needed a fitness value, and that's simple.
Then I needed training images for the VQ to come back, and I added
a facility to fetch them.
Now I am also computing per-centroid utilities, and that is sign it
is time to organize all the message-passing I want between children
and parent.